:recycle: refactor: EmailService 및 EmailDataFetchService 리팩토링
- 레포지토리를 리팩토링하면서 EmailDataFetchService에서 호출하는 메서드 또한 수정

- 이메일 알림에 필요한 데이터를 조회한다는 의미로 클래스 이름을 'EmailDataFetchService'로 변경

- EmailService의 매개변수 타입을 List에서 Set으로 수정
